terminal-setup: Fix annoying behavior related to 'project.el'
Remove 'project' awareness from 'cwd', which obviates the
'maybe-in-project' macro (now removed). The old behavior of 'cwd' was
to return the root directory of the project (if called from a
project), which resulted in the annoying behavior of a vterm session
opening in the project root instead of in your intended
subdirectory (when trying to open a session from a subdirectory of a
project).

There may be a way to make the commands given by 'terminal-setup' more
friendly by including some other kind of integration with
'project.el', but I'll leave this for a later time.