mail-setup: Stop EBDB from auto-querying
The automatic query dialogue upon viewing a message was getting a bit
tiresome. So, disable it. Now, when you want to add a new entry to the
address book, use the "; :" keybinding. This brings up the dialogue
that used to come up automatically.

I like the "opt-in" method here, since much of the mail I receive is
from senders who I don't reply to, and hence don't need in my address
book.