functions-setup: Decentralize, and remove
The keys-setup and package-setup refactors (which also used the term
"Decentralize" in their commit messages) ended up moving a bunch of
functions out of functions-setup and into more logical locations. Only
a few functions remained, most of which I hadn't used in a long time.
Remove those; move the 2 that I want to keep over to appearance-setup
where they fit nicely; and, with functions-setup now empty, remove it.